Refactor binary transport into focused modules
